+++ CHANGELOG.rst
@@ -18,12 +18,13 @@ v4.35.0 (2024-12-??)
 Added
 ^^^^^
 - Support for ``print config`` argument to reuse the name of the config argument
-  by using ``%s`` (`#??? <https://github.com/omni-us/jsonargparse/pull/???>`__).
+  by using ``%s`` (`#630 <https://github.com/omni-us/jsonargparse/pull/630>`__).
 
 Deprecated
 ^^^^^^^^^^
 - From v5.0.0 the print config argument will by default reuse the name of the
-  config argument as ``--print_%s`` instead of being always ``--print_config``.
+  config argument as ``--print_%s`` instead of being always ``--print_config``
+  (`#630 <https://github.com/omni-us/jsonargparse/pull/630>`__).
 
 
 v4.34.1 (2024-11-??)